+++ lib/start.dart
@@ -1,9 +1,8 @@
 library start;
 
-import 'dart:io' hide Socket, Cookie;
+import 'dart:io' hide Socket;
 import 'dart:async';
 import 'dart:json' as Json;
-import 'src/cookie.dart';
 
 part 'src/route.dart';
 part 'src/request.dart';
@@ -13,6 +12,6 @@ part 'src/socket.dart';
 part 'src/server.dart';
 part 'src/mime_types.dart';
 
-Future<Server> start({ view, String public: 'web', String host: '127.0.0.1', int port: 80 }) {
-  return new Server(view, public).listen(host, port);
+Future<Server> start({ String public: 'web', String host: '127.0.0.1', int port: 80 }) {
+  return new Server(public).listen(host, port);
 }